cast to long once more

+++ pcdet/models/dense_heads/detr/detr.py
@@ -108,7 +108,7 @@ class SetCriterion(nn.Module):
         src_logits = outputs['cls_preds']
 
         idx = self._get_src_permutation_idx(indices)
-        target_classes_o = torch.cat([t["labels"][J] for t, (_, J) in zip(targets, indices)])
+        target_classes_o = torch.cat([t["labels"][J] for t, (_, J) in zip(targets, indices)]).long()
         target_classes = torch.full(src_logits.shape[:2], self.num_classes,
                                     dtype=torch.int64, device=src_logits.device)
         target_classes[idx] = target_classes_o
